[tests] Reodered phone and rest token tests
- test_phone_token > test_phone_verification
- tests for auth token validation moved to test_rest_token
